+++ src/main/java/com/vidaloca/skibidi/event/repository/EventRepository.java
@@ -3,5 +3,5 @@ package com.vidaloca.skibidi.event.repository;
 import com.vidaloca.skibidi.event.model.Event;
 import org.springframework.data.repository.CrudRepository;
 
-public interface EventRepository extends CrudRepository<Event,Integer> {
+public interface EventRepository extends CrudRepository<Event,Long> {
 }
